yay! colored logs

+++ src/dtos/http_responses.go
@@ -3,7 +3,7 @@ package dtos
 import "onepixel_backend/src/models"
 
 type UserResponse struct {
-	ID    uint    `json:"id" example:"1"`
+	ID    uint64  `json:"id" example:"1"`
 	Email string  `json:"email" example:"user@test.com"`
 	Token *string `json:"token" example:"<JWT_TOKEN>"`
 }
